feat: disable variable option when there are no variables, break down components
